refactor(twitter): move token store out of @umwelten/habitat into examples/twitter-habitat
Keep the generic agent-container library clean: the Twitter habitat is a
self-contained deployable example agent (like examples/jeeves-bot and
examples/twitter-mcp), so its deep modules belong in examples/, not in the
published @umwelten/habitat package.

- Move x-oauth.ts, token-store.ts and their tests to examples/twitter-habitat/src/
- Revert the @umwelten/habitat index.ts exports (now identical to main)
- bootstrap-oauth.ts imports the local ./src/x-oauth.js
- Widen the root vitest include with examples/twitter-habitat/**/*.test.ts so the
  token-store unit tests still run under `pnpm test:run` (and add an explicit
  node_modules exclude). Full suite green: 1458 passed.
