+++ vitest.config.ts
@@ -2,8 +2,17 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
 	test: {
-		include: ["packages/*/src/**/*.test.ts", "packages/*/src/**/*.test.tsx"],
-		exclude: ["packages/*/src/**/*.integration.test.ts"],
+		include: [
+			"packages/*/src/**/*.test.ts",
+			"packages/*/src/**/*.test.tsx",
+			// Deployable example agents that ship their own tested deep modules
+			// (kept out of the library packages, but still run in the default suite).
+			"examples/twitter-habitat/**/*.test.ts",
+		],
+		exclude: [
+			"packages/*/src/**/*.integration.test.ts",
+			"**/node_modules/**",
+		],
 		environment: "node",
 		globals: true,
 		coverage: {
